requirements.txt
mysql-connector-python==8.0.32
web3==6.0.0
requests==2.28.2
blocksmith==1.2.1
cryptocode==0.1
python-dotenv==1.0.0
Flask==2.3.2
Flask-Bcrypt==1.0.1
Flask-Mail==0.9.1
Flask-Cors==4.0.0
PyJWT==2.8.0
hexbytes==0.3.1
solathon==0.2.0
